requirements.txt
Django
djangocms-plus